refactor: move id-validation fucntion to utils
