+++ openstates/utils/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 from .generic import (
     _make_pseudo_id,
+    is_valid_uuid,
     get_pseudo_id,
     makedirs,
     JSONEncoderPlus,
